+++ src/main/java/com/alvarx4z/woja/infrastructure/controller/expansion/ExpansionControllerAPI.java
@@ -1,5 +1,6 @@
 package com.alvarx4z.woja.infrastructure.controller.expansion;
 
+import com.alvarx4z.woja.infrastructure.controller.BaseControllerAPI;
 import com.alvarx4z.woja.infrastructure.controller.expansion.dto.response.ExpansionsResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -9,7 +10,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestHeader;
 
 @Tag(name = "Expansions", description = "Expansions API")
 public interface ExpansionControllerAPI extends BaseControllerAPI {
